fix refresh correctly token

+++ frontend/lib/features/token/domain/use_cases/refresh_token_use_case.dart
@@ -1,33 +1,38 @@
-import 'package:cookify/core/domain/use_cases/results/result.dart';
+import 'package:cookify/core/domain/my_either/my_either.dart';
+import 'package:cookify/features/token/domain/failures/token_failures.dart';
 import 'package:cookify/features/token/domain/payloads/refresh_token_payload.dart';
-import 'package:cookify/features/token/domain/payloads/save_token_payload.dart';
+import 'package:cookify/features/token/domain/payloads/set_token_payload.dart';
 import 'package:cookify/features/token/domain/repositories/token_repository.dart';
+import 'package:fpdart/fpdart.dart';
 
 class RefreshTokenUseCase {
   RefreshTokenUseCase(this._repository);
 
   final TokenRepository _repository;
 
-  Future<Result<void>> call() async {
-    try {
-      return Result.success(await execute());
-    } on Exception catch (e) {
-      return Result.failure(e);
-    }
-  }
+  Future<MyEither<void>> call() async {
+    final tokenResult = await _repository.getToken();
 
-  Future<void> execute() async {
-    final token = (await _repository.getToken())!;
+    return tokenResult.fold((failure) => Left(failure), (token) async {
+      if (token == null) {
+        return const Left(NotFoundTokenFailure());
+      }
 
-    try {
-      final newToken = await _repository.refreshToken(
+      final newTokenResult = await _repository.refreshToken(
         RefreshTokenPayload(refreshToken: token.refreshToken),
       );
 
-      await _repository.saveToken(SaveTokenPayload(token: newToken));
-    } on Exception {
-      await _repository.deleteToken();
-      rethrow;
-    }
+      return newTokenResult.fold(
+        (failure) async {
+          await _repository.markTokenAsInvalid();
+          await _repository.deleteToken();
+
+          return Left(failure);
+        },
+        (newToken) {
+          return _repository.setToken(SetTokenPayload(token: newToken));
+        },
+      );
+    });
   }
 }
